Made the webpage work properly, refactored code and added comments
When an employee's gallery card is clicked, their modal card is displayed on the screen.  When the modal window x button is clicked, the modal window is closed.  The process is repeatable with any other card clicks.  Also, the number of loops for the for loops determines how much data is retrieved from the server.